+++ src/ascat/__init__.py
@@ -4,6 +4,3 @@ try:
     __version__ = pkg_resources.get_distribution(__name__).version
 except:
     __version__ = 'unknown'
-
-from ascat.read_native.cdr import *
-from ascat.h_saf import *