+++ src/ascat/h_saf.py
@@ -47,9 +47,6 @@ import ascat.read_native.bufr as bufr_reader
 from ascat.read_native.bufr import AscatL2SsmBufr
 from ascat.read_native.cdr import AscatNc
 
-if sys.version_info < (3, 0):
-    range = xrange
-
 
 class H08Single(ImageBase):