+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog
 =========
 
+Version 1.2.0
+=============
+
+- Add download interface for H SAF FTP and EUMETSAT Data Store
+- Move CI to Github actions (Ubuntu and Windows CI)
+
 Version 1.1.2
 =============